EEBus meter: fix monitoring of power consumption
